feature useParams to get note id in Note page when navigate from note link and graphView
