+++ src/app/loginSlice.ts
@@ -1,6 +1,7 @@
 import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { RootState } from './store';
 import { auth, firestore } from '../utils/firebase';
+import { NoteType } from './types';
 
 interface CustomUser {
   uid: string;
@@ -11,23 +12,8 @@ interface Profile {
   email: string;
   name: string;
   id: string;
-  notes: Note[];
+  notes: NoteType[];
   isLogin: boolean;
-}
-
-interface linkObj {
-  id: string;
-  title: string;
-}
-
-interface Note {
-  id: string,
-  title: string,
-  content: string,
-  category: string,
-  link_notes: linkObj[] | [],
-  create_time: string,
-  edit_time: string,
 }
 
 const initialState = {
@@ -35,7 +21,7 @@ const initialState = {
     email: '',
     name: '',
     id: '',
-    notes: [] as Note[],
+    notes: [] as NoteType[],
     isLogin: false,
   }
 }
@@ -106,8 +92,8 @@ const loginSlice = createSlice({
   name : 'user',
   initialState : initialState,
   reducers:{
-    addNote: (state, action: PayloadAction<Note>) => {
-      state.profile.notes.push(action.payload as Note);
+    addNote: (state, action: PayloadAction<NoteType>) => {
+      state.profile.notes.push(action.payload as NoteType);
       const id = state.profile.id
       const data = {
         email: state.profile.email,
