+++ src/pages/Landing/Landing.tsx
@@ -10,4 +10,3 @@ const Landing = () => {
 };
 
 export default Landing;
-//TODO: localstorage login
